Refactoring. Разделил логику по разным файлам
